0.4.1: money is decimal in the model, because the product's own contract says so. (#20)
`money_is_never_stored_as_float` requires every column named like money to be
exact and base-10 in the WAREHOUSE. The semantic model served those same
columns as `double`, widening them back to a binary float on the way to a BI
client -- which is the rule's spirit broken one layer out from where it is
enforced.

IT CHANGES NOTHING AGAINST THIS FAMILY'S EMULATOR, checked rather than
assumed: `daxDataType` maps decimal and double alike to DOUBLE, and
`summableType` accepts both, so the evaluator sums float64 whatever a column
declares. On real Fabric the engine honours the declared type. That gap is
precisely why this needed a test -- nothing measurable here would ever catch
it, which is how it shipped in 0.4.0 in the first place.

The test reads the money-like fragments FROM THE CONTRACT rather than
restating them. The contract matches by name on purpose ("the next money
column somebody adds is exactly the one that would slip through"), so a second
copy of that list would be the drift this repo exists to prevent -- the same
reason COLUMNS is checked against schema.yml. It also refuses to pass
vacuously: if no modelled column matches the fragments, the test fails saying
so.

Checked in both directions: reverting revenue_usd to `double` fails by name,
quoting the fragments and citing the contract file.

Consumers see no numeric change. A leaf on 0.4.0 is not wrong where it is
measured; it is wrong where it would be deployed.

+++ src/contoso_product/semantic.py
@@ -45,9 +45,23 @@ TABLE = "fct_revenue_summary"
 # Types are declared here and only here: the yml carries quality contracts,
 # not storage types, and a semantic model needs both.
 COLUMNS: tuple[tuple[str, str], ...] = (
-    ("revenue_usd", "double"),
+    # MONEY IS `decimal`, NOT `double`, and the product's own contract says so:
+    # `money_is_never_stored_as_float` requires every column named like money to
+    # be an exact base-10 type in the warehouse, and it would be strange for the
+    # model SERVING those columns to widen them back to a binary float on the
+    # way out. That contract exists because a single `cast(null as float)` in a
+    # UNION once demoted every real price silently -- every row still held the
+    # right number and the column had stopped being money.
+    #
+    # It changes nothing against this family's emulator, whose evaluator sums in
+    # float64 whatever a column declares -- checked, not assumed
+    # (`summableType` and `daxDataType` both accept decimal and both land on
+    # DOUBLE). On real Fabric the engine honours the declared type, so this is
+    # the difference between a model that is correct where it is measured and
+    # one that is correct where it is deployed.
+    ("revenue_usd", "decimal"),
     ("sale_lines", "int64"),
-    ("cancelled_revenue_usd", "double"),
+    ("cancelled_revenue_usd", "decimal"),
     ("fiscal_year_label", "string"),
     ("customer_segment", "string"),
     ("product_segment", "string"),
